minor code cleanup.

+++ takehiro.c
@@ -638,12 +638,12 @@ void best_scalefac_store(lame_internal_flags *gfc,int gr, int ch,
 {
     /* use scalefac_scale if we can */
     gr_info *gi = &l3_side->gr[gr].ch[ch].tt;
-    u_int sfb,i,j,j2,l,start,end;
+    int sfb,i,j,j2,l,start,end;
 
     /* remove scalefacs from bands with ix=0.  This idea comes
      * from the AAC ISO docs.  added mt 3/00 */
     /* check if l3_enc=0 */
-    for ( sfb = 0; sfb < (u_int)gi->sfb_lmax; sfb++ ) {
+    for ( sfb = 0; sfb < gi->sfb_lmax; sfb++ ) {
       if (scalefac[gr][ch].l[sfb]>0) { 
 	start = gfc->scalefac_band.l[ sfb ];
 	end   = gfc->scalefac_band.l[ sfb+1 ];
@@ -669,7 +669,7 @@ void best_scalefac_store(lame_internal_flags *gfc,int gr, int ch,
     gi->part2_3_length -= gi->part2_length;
     if (!gi->scalefac_scale && !gi->preflag) {
 	int b, s = 0;
-	for (sfb = 0; sfb < (u_int)gi->sfb_lmax; sfb++) {
+	for (sfb = 0; sfb < gi->sfb_lmax; sfb++) {
 	    s |= scalefac[gr][ch].l[sfb];
 	}
 
@@ -680,7 +680,7 @@ void best_scalefac_store(lame_internal_flags *gfc,int gr, int ch,
 	}
 
 	if (!(s & 1) && s != 0) {
-	    for (sfb = 0; sfb < (u_int)gi->sfb_lmax; sfb++) {
+	    for (sfb = 0; sfb < gi->sfb_lmax; sfb++) {
 		scalefac[gr][ch].l[sfb] /= 2;
 	    }
 	    for (sfb = gi->sfb_smax; sfb < SBPSY_s; sfb++) {
